+++ Realm/Tests/ObjectInterfaceTests.m
@@ -17,60 +17,54 @@
 ////////////////////////////////////////////////////////////////////////////
 
 #import "RLMTestCase.h"
-#import "RLMTestObjects.h"
 
-//
-// for custom accessor test
-//
-@interface CustomAccessors : RLMObject
-@property (getter = getThatName) NSString * name;
-@property (setter = setTheInt:) int age;
+#pragma mark - Test Objects
+
+#pragma mark CustomAccessorsObject
+
+@interface CustomAccessorsObject : RLMObject
+@property (getter = getThatName) NSString *name;
+@property (setter = setTheInt:)  int age;
 @end
 
-@implementation CustomAccessors
+@implementation CustomAccessorsObject
 @end
 
+#pragma mark InvalidSubclassObject
 
-//
-// for subclass test
-//
-@interface InvalidSubclassObject : RLMTestObject
+@interface InvalidSubclassObject : StringObject
 @property NSString *invalid;
 @end
 
 @implementation InvalidSubclassObject
 @end
 
+#pragma mark BaseClassStringObject
 
-//
-// for class extension test
-//
-@interface BaseClassTestObject : RLMObject
+@interface BaseClassStringObject : RLMObject
 @property NSInteger intCol;
 @end
 
-// Class extension, adding one more column
-@interface BaseClassTestObject ()
-@property (nonatomic, copy) NSString *stringCol;
+@interface BaseClassStringObject ()
+@property NSString *stringCol;
 @end
 
-@implementation BaseClassTestObject
+@implementation BaseClassStringObject
 @end
 
-
+#pragma mark - Tests
 
 @interface ObjectInterfaceTests : RLMTestCase
 @end
 
 @implementation ObjectInterfaceTests
 
-
-- (void)testCustomAccessors
+- (void)testCustomAccessorsObject
 {
     RLMRealm *realm = [RLMRealm defaultRealm];
     
     [realm beginWriteTransaction];
-    CustomAccessors *ca = [CustomAccessors createInRealm:realm withObject:@[@"name", @2]];
+    CustomAccessorsObject *ca = [CustomAccessorsObject createInRealm:realm withObject:@[@"name", @2]];
     XCTAssertEqualObjects([ca getThatName], @"name", @"name property should be name.");
     
     [ca setTheInt:99];
@@ -94,14 +88,13 @@
     RLMRealm *realm = [RLMRealm defaultRealm];
     
     [realm beginWriteTransaction];
-    BaseClassTestObject *bObject = [[BaseClassTestObject alloc ] init];
+    BaseClassStringObject *bObject = [[BaseClassStringObject alloc ] init];
     bObject.intCol = 1;
     bObject.stringCol = @"stringVal";
     [realm addObject:bObject];
     [realm commitWriteTransaction];
     
-    
-    BaseClassTestObject *objectFromRealm = [BaseClassTestObject allObjects][0];
+    BaseClassStringObject *objectFromRealm = [BaseClassStringObject allObjects][0];
     XCTAssertEqual(1, objectFromRealm.intCol, @"Should be 1");
     XCTAssertEqualObjects(@"stringVal", objectFromRealm.stringCol, @"Should be stringVal");
 }