+++ CHANGELOG.md
@@ -5,12 +5,18 @@
 
 ### API breaking changes
 
-* Rename schemaForObject: to schemaForClassName: on RLMSchema 
+* Rename schemaForObject: to schemaForClassName: on RLMSchema
+* Removed -objects:where: and -objects:orderedBy:where: from RLMRealm
+* Removed -indexOfObjectWhere:, -objectsWhere: and -objectsOrderedBy:where: from RLMArray
+* Removed +objectsWhere: and +objectsOrderedBy:where: from RLMObject
 
 ### Enhancements
 
 * Added support for RLMArrayProperties.
 * Added support for creating in-memory default realm.
+* Added -objectsWithClassName:predicateFormat: and -objectsWithClassName:predicate: to RLMRealm
+* Added -indexOfObjectWithPredicateFormat:, -indexOfObjectWithPredicate:, -objectsWithPredicateFormat:, -objectsWithPredicate: and -arraySortedByProperty:ascending: to RLMArray
+* Added +objectsWithPredicateFormat: and +objectsWithPredicate: to RLMObject
 
 ### Bugfixes